Add an option to save the toolchain tarballs.
Adds the --save-downloads-dir option to download_toolchains.py and perculates
it through to sync_tgz.  If this option is not set, behvaiour of download_
toochains.py is exactly as it was before.

This option is used by the SDK to create NSIS manifests from the table of
contents in the toolchain tarballs.

Review URL: http://codereview.chromium.org/7827036
Patch from Dave Springer <dspringer@chromium.org>.

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@6622 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2

Former-commit-id: d70e11831294902284e0de4744a91c8e134ea520

+++ build/sync_tgz.py
@@ -76,7 +76,8 @@ class HashError(Exception):
 
 
 def SyncTgz(url, target, maindir='sdk',
-            username=None, password=None, verbose=True, hash=None):
+            username=None, password=None, verbose=True, hash=None,
+            save_path=None):
   """Download a file from a remote server.
 
   Args:
@@ -86,9 +87,14 @@ def SyncTgz(url, target, maindir='sdk',
     password: Optional password for download (ignored if no username).
     verbose: Flag indicating if status shut be printed.
     hash: sha1 of expected download (don't check if hash is None).
+    save_path: Optional place to save the downloaded tarball.  If this is None,
+        then the tarball is downloaded to a temporary file which is deleted.
   """
   shutil.rmtree(target, True)
+  os.makedirs(target)
   tgz_filename = target + '/.tgz'
+  if save_path:
+    tgz_filename = save_path
 
   if verbose:
     print 'Downloading %s to %s...' % (url, tgz_filename)
@@ -176,7 +182,8 @@ def SyncTgz(url, target, maindir='sdk',
         print m.name
       tgz.extract(m, target)
     tgz.close()
-  os.remove(tgz_filename)
+  if save_path is None:
+    os.remove(tgz_filename)
 
   if verbose:
     print 'Update complete.'
